# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Dekopen API
  version: 1.0.0
  description: Authenticated tenant and engine API boundary.
paths:
...
  /api/v1/documents/artifacts/:
    post:
      operationId: documentary_generate_artifact
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ArtifactRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ArtifactRequestRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/documents/artifacts/{artifact_id}/access/:
    post:
      operationId: documentary_artifact_access
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: artifact_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignedAccessResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/documents/projects/{project_id}/freeze/:
    post:
      operationId: documentary_freeze_revision_a
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreezeRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FreezeRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FreezeRequestRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreezeResponse'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreezeResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
  /api/v1/documents/projects/{project_id}/inputs/:
    get:
      operationId: documentary_prepare_inputs
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentaryPreparationResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
    put:
      operationId: documentary_save_inputs
      parameters:
      - in: header
        name: X-Organization-ID
        schema:
          type: string
          format: uuid
        description: Active organization; validated against the JWT user's active
          memberships.
      - in: path
        name: project_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentaryInputsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DocumentaryInputsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentaryInputsRequest'
        required: true
      security:
      - SupabaseBearer: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentaryInputsResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: ''
...
components:
  schemas:
...
    AccessorySchedule:
      type: object
      properties:
        schema_version:
          type: integer
          maximum: 1
          minimum: 1
          default: 1
        coverage:
          $ref: '#/components/schemas/CoverageEnum'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccessoryLine'
      required:
      - coverage
      - items
...
    ArtifactRequestRequest:
      type: object
      properties:
        document_type:
          $ref: '#/components/schemas/DocumentTypeEnum'
        format:
          $ref: '#/components/schemas/FormatEnum'
        project_version_id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
          nullable: true
      required:
      - document_type
      - format
      - project_version_id
    ArtifactResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        artifact_scope:
          $ref: '#/components/schemas/ArtifactScopeEnum'
        artifact_scope_id:
          type: string
          format: uuid
        document_type:
          type: string
        format:
          type: string
        bom_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
        file_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        byte_size:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
      - artifact_scope
      - artifact_scope_id
      - bom_hash
      - byte_size
      - created_at
      - document_type
      - file_sha256
      - format
      - id
    ArtifactScopeEnum:
      enum:
      - PROJECT_REVISION
      - ORDER
      type: string
      description: |-
        * `PROJECT_REVISION` - PROJECT_REVISION
        * `ORDER` - ORDER
...
    DocumentTypeEnum:
      enum:
      - DOC-01
      - DOC-02
      - DOC-03
      - DOC-04
      - DOC-05
      - DOC-06
      - DOC-07
      type: string
      description: |-
        * `DOC-01` - DOC-01
        * `DOC-02` - DOC-02
        * `DOC-03` - DOC-03
        * `DOC-04` - DOC-04
        * `DOC-05` - DOC-05
        * `DOC-06` - DOC-06
        * `DOC-07` - DOC-07
    DocumentaryInputsRequest:
      type: object
      properties:
        payment_terms:
          type: string
          minLength: 1
          maxLength: 2000
        quotation_valid_until:
          type: string
          format: date
        positions:
          type: array
          items:
            $ref: '#/components/schemas/PositionDocumentaryInputRequest'
      required:
      - payment_terms
      - positions
      - quotation_valid_until
    DocumentaryInputsResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        positions_saved:
          type: integer
      required:
      - positions_saved
      - project_id
...
    DocumentaryPreparationPosition:
      type: object
      properties:
        position_id:
          type: string
          format: uuid
        location_tag:
          type: string
          maxLength: 100
        manufacturing_placement_policy_id:
          type: string
          format: uuid
          nullable: true
        handle_requirement_policy_id:
          type: string
          format: uuid
          nullable: true
        reinforcement_cut_policy_id:
          type: string
          format: uuid
          nullable: true
        workshop_annotations:
          type: array
          items:
            $ref: '#/components/schemas/WorkshopAnnotation'
        structural_inputs:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryStructuralInput'
        glass_polishing:
          type: array
          items:
            $ref: '#/components/schemas/GlassPolishing'
        handle_intents:
          type: array
          items:
            $ref: '#/components/schemas/HandleIntent'
        accessory_schedule:
          $ref: '#/components/schemas/AccessorySchedule'
        legacy_handle_migration_confirmed:
          type: boolean
          default: false
        system_name:
          type: string
        placement_options:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryPolicyOption'
        handle_options:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryPolicyOption'
        reinforcement_options:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryPolicyOption'
      required:
      - accessory_schedule
      - glass_polishing
      - handle_intents
      - handle_options
      - handle_requirement_policy_id
      - location_tag
      - manufacturing_placement_policy_id
      - placement_options
      - position_id
      - reinforcement_cut_policy_id
      - reinforcement_options
      - structural_inputs
      - system_name
      - workshop_annotations
    DocumentaryPreparationResponse:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        revision_code:
          type: string
          pattern: ^REV-[A-Z]+$
        payment_terms:
          type: string
        quotation_valid_until:
          type: string
          format: date
          nullable: true
        positions:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryPreparationPosition'
      required:
      - payment_terms
      - positions
      - project_id
      - quotation_valid_until
      - revision_code
    DocumentaryStructuralInput:
      type: object
      properties:
        target_id:
          type: string
          maxLength: 200
        required_ix_cm4:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,4})?$
          nullable: true
        structural_basis:
          type: string
          nullable: true
          maxLength: 1000
      required:
      - target_id
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        detail:
          type: string
        required_aal:
          $ref: '#/components/schemas/RequiredAalEnum'
      required:
      - code
      - detail
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/Membership'
      required:
      - error
...
    FormatEnum:
      enum:
      - PDF
      - XLSX
      type: string
      description: |-
        * `PDF` - PDF
        * `XLSX` - XLSX
    FreezeRequestRequest:
      type: object
      properties:
        pricing_operation_id:
          type: string
          format: uuid
        confirmed:
          type: boolean
      required:
      - confirmed
      - pricing_operation_id
    FreezeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pricing_operation_id:
          type: string
          format: uuid
        created:
          type: boolean
        revision_code:
          type: string
          pattern: ^REV-[A-Z]+$
        bom_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
        snapshot_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
        production_allowed:
          type: boolean
        documentary_complete:
          type: boolean
        emitted_at:
          type: string
          format: date-time
        purchase_projection_hash:
          type: string
          pattern: ^[0-9a-f]{64}$
      required:
      - bom_hash
      - created
      - documentary_complete
      - emitted_at
      - id
      - pricing_operation_id
      - production_allowed
      - revision_code
      - snapshot_sha256
...
    GlassPolishing:
      type: object
      properties:
        schema_version:
          type: integer
          maximum: 1
          minimum: 1
          default: 1
        bay_id:
          type: string
          maxLength: 200
        leaf_id:
          type: string
          nullable: true
          maxLength: 200
        edges:
          $ref: '#/components/schemas/PolishingEdges'
      required:
      - bay_id
      - edges
...
    HandleIntent:
      type: object
      properties:
        schema_version:
          type: integer
          maximum: 1
          minimum: 1
          default: 1
        bay_id:
          type: string
          maxLength: 200
        leaf_id:
          type: string
          nullable: true
          maxLength: 200
        handle_domain_slot:
          type: string
          maxLength: 100
        requested_height_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,4})?$
        vertical_reference:
          $ref: '#/components/schemas/VerticalReferenceEnum'
      required:
      - bay_id
      - handle_domain_slot
      - requested_height_mm
      - vertical_reference
...
    Membership:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
        organization_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
      required:
      - organization_id
      - organization_name
      - role
...
    NullEnum:
      enum:
      - null
...
    PositionDocumentaryInputRequest:
      type: object
      properties:
        position_id:
          type: string
          format: uuid
        location_tag:
          type: string
          minLength: 1
          maxLength: 100
        manufacturing_placement_policy_id:
          type: string
          format: uuid
        handle_requirement_policy_id:
          type: string
          format: uuid
        reinforcement_cut_policy_id:
          type: string
          format: uuid
        workshop_annotations:
          type: array
          items:
            $ref: '#/components/schemas/WorkshopAnnotationRequest'
        structural_inputs:
          type: array
          items:
            $ref: '#/components/schemas/DocumentaryStructuralInputRequest'
        glass_polishing:
          type: array
          items:
            $ref: '#/components/schemas/GlassPolishingRequest'
        handle_intents:
          type: array
          items:
            $ref: '#/components/schemas/HandleIntentRequest'
        accessory_schedule:
          $ref: '#/components/schemas/AccessoryScheduleRequest'
        legacy_handle_migration_confirmed:
          type: boolean
          default: false
      required:
      - accessory_schedule
      - glass_polishing
      - handle_intents
      - handle_requirement_policy_id
      - location_tag
      - manufacturing_placement_policy_id
      - position_id
      - reinforcement_cut_policy_id
      - structural_inputs
      - workshop_annotations
...
    RequiredAalEnum:
      enum:
      - aal2
      type: string
      description: '* `aal2` - aal2'
    RoleEnum:
      enum:
      - OWNER
      - ESTIMATOR
      - WORKSHOP_MANAGER
      - INSTALLER
      type: string
      description: |-
        * `OWNER` - OWNER
        * `ESTIMATOR` - ESTIMATOR
        * `WORKSHOP_MANAGER` - WORKSHOP_MANAGER
        * `INSTALLER` - INSTALLER
...
    SignedAccessResponse:
      type: object
      properties:
        artifact_id:
          type: string
          format: uuid
        signed_url:
          type: string
          format: uri
        expires_in:
          type: integer
          maximum: 3600
          minimum: 3600
      required:
      - artifact_id
      - expires_in
      - signed_url
...
    WorkshopAnnotation:
      type: object
      properties:
        bay_id:
          type: string
          maxLength: 200
        leaf_id:
          type: string
          nullable: true
          maxLength: 200
        bottom_drain_holes_mm:
          type: array
          items:
            type: string
            format: decimal
            pattern: ^-?\d{0,10}(?:\.\d{0,4})?$
          nullable: true
        closing_points_perimeter_mm:
          type: array
          items:
            type: string
            format: decimal
            pattern: ^-?\d{0,10}(?:\.\d{0,4})?$
          nullable: true
        continuous_width_mm:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,4})?$
          nullable: true
        finish_class:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/FinishClassEnum'
          - $ref: '#/components/schemas/NullEnum'
        has_coupler:
          type: boolean
          nullable: true
      required:
      - bay_id
...
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT